fix: refactor downloadEncryptedFile sample into its own file
